requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Database
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.25
alembic>=1.13.1
pgvector>=0.2.5

# Redis
redis>=5.0.1

# AI/ML
google-genai>=1.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-core>=0.1.0
langsmith>=0.7.0

# HTTP Client
httpx>=0.26.0

# Logging
structlog>=24.1.0

# Utilities
python-dotenv>=1.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
ruff>=0.2.0

# Type stubs
types-redis>=4.6.0
